Added csv parse functionality

+++ client/app/ParseTime.js
@@ -5,10 +5,10 @@ export function parseTime(timePeriod, time) {
       options = {timeZone: 'UTC', hour: 'numeric', hour12: true}
       return currDate = new Date(time).toLocaleString('en-US', options)
     case 1:
-      options = {timeZone: 'UTC', weekday: 'short', month: 'short', day: 'numeric'}
+      options = {timeZone: 'UTC', weekday: 'short', month: 'numeric', day: 'numeric'}
       return currDate = new Date(time).toLocaleDateString('en-US', options)
     case 2:
-      options = {timeZone: 'UTC', weekday: 'short', month: 'short', day: 'numeric'}
+      options = {timeZone: 'UTC', weekday: 'short', month: 'numeric', day: 'numeric'}
       return currDate = new Date(time).toLocaleDateString('en-US', options)
     case 3:
       const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
